// src/data/api/models/get-sensor-data-query-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * lida-api
 * This is a REST API utilising the CQRS design pattern being exposed via Google Clouds serverless functions.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 *
 * @export
 * @interface GetSensorDataQueryDto
 */
export interface GetSensorDataQueryDto {
  /**
   *
   * @type {string}
   * @memberof GetSensorDataQueryDto
   */
  sensorName: any
  /**
   *
   * @type {number}
   * @memberof GetSensorDataQueryDto
   */
  value: any
  /**
   *
   * @type {number}
   * @memberof GetSensorDataQueryDto
   */
  timeStamp: any
}